Upgraded GLM to the latest version
With this change the repository now include a submodule that point to
the version of the glm that's currently exposed. The library has also
been updated to cover compilation of the template so that it can be
used in projects easily.

The internal version has been updated to reflect the changes.